Fixes FOC on opening slides for presentation
Adjusts the proportion of the banner on the first two slides of the
template in order to avoid de Flash of Content when changing slides.

As the first banner is bigger than the second one, the second one
"blinked" when changing slides. This was fixed by adjusting the
proportion of the second banner to be the same as the first one.

The aforementioned change maintains the aspect ratio of the images
and the text, so the slides are still visually appealing and similar.